fix for JSON.parse() error when using jQuery 1.4+
in jQuery 1.4 and up the $.ajax() call will now automatically convert
the "data" returned into a JS Object when using "type:json". The plugin
was always calling JSON.parse(remoteData) because older versions of
jQuery always returned a json string.

Added a check to make sure that type == json AND remoteData is a
string before trying to parse remoteData. This will allow the correct
parsing to happen for pre1.4 jQuery, while skipping this step when
remoteData is already an Object (the case in jQuery 1.4+)

Verified that this works nicely with jQuery 1.3.2 and jQuery 1.4.2